+++ apps/web/src/components/Builders/SmartWallet/onboarding.png
